@@ -1,6 +1,6 @@
 module proto
 
-go 1.23
+go 1.25.5
 
 require (
 	google.golang.org/grpc v1.69.2
