@@ -1,11 +1,10 @@
 package commands
 
 import (
+	"crypto_vault_service/internal/domain/crypto"
 	"crypto_vault_service/internal/infrastructure/cryptography"
-	"crypto_vault_service/internal/infrastructure/logger"
-	"crypto_vault_service/internal/infrastructure/settings"
+	"crypto_vault_service/internal/pkg/logger"
 	"fmt"
-	"log"
 	"os"
 	"path/filepath"
 
@@ -15,128 +14,157 @@ import (
 
 // AESCommandHandler encapsulates logic for handling AES operations via CLI.
 type AESCommandHandler struct {
-	aesProcessor cryptography.AESProcessor
-	Logger       logger.Logger
+	aesProcessor crypto.AESProcessor
+	logger       logger.Logger
 }
 
 // NewAESCommandHandler initializes and returns an AESCommandHandler instance with
 // configured logger and AES processor.
-func NewAESCommandHandler() *AESCommandHandler {
-	loggerSettings := &settings.LoggerSettings{
-		LogLevel: "info",
-		LogType:  "console",
-		FilePath: "",
+func NewAESCommandHandler() (*AESCommandHandler, error) {
+	loggerInstance, err := setupLogger()
+	if err != nil {
+		return nil, fmt.Errorf("failed to setup logger: %w", err)
 	}
 
-	logger, err := logger.GetLogger(loggerSettings)
+	aesProcessor, err := cryptography.NewAESProcessor(loggerInstance)
 	if err != nil {
-		log.Panicf("Error creating logger: %v", err)
-		return nil
-	}
-
-	aesProcessor, err := cryptography.NewAESProcessor(logger)
-	if err != nil {
-		log.Panicf("%v\n", err)
-		return nil
+		return nil, fmt.Errorf("failed to create AES processor: %w", err)
 	}
 
 	return &AESCommandHandler{
 		aesProcessor: aesProcessor,
-		Logger:       logger,
-	}
+		logger:       loggerInstance,
+	}, nil
 }
 
 // GenerateAESKeysCmd generates AES key pairs and persists those in a selected directory
 func (commandHandler *AESCommandHandler) GenerateAESKeysCmd(cmd *cobra.Command, _ []string) {
-	keySize, _ := cmd.Flags().GetInt("key-size")
-	keyDir, _ := cmd.Flags().GetString("key-dir")
+	keySize, err := cmd.Flags().GetInt("key-size")
+	if err != nil {
+		commandHandler.logger.Error("invalid key-size flag ", err)
+		return
+	}
+
+	keyDir, err := cmd.Flags().GetString("key-dir")
+	if err != nil {
+		commandHandler.logger.Error("invalid key-dir flag ", err)
+		return
+	}
 
 	uniqueID := uuid.New()
 
 	secretKey, err := commandHandler.aesProcessor.GenerateKey(keySize)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	keyFilePath := filepath.Join(keyDir, fmt.Sprintf("%s-symmetric-key.bin", uniqueID))
 	err = os.WriteFile(keyFilePath, secretKey, 0600)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
-	commandHandler.Logger.Info(fmt.Sprintf("AES key saved to %s", keyFilePath))
+	commandHandler.logger.Info("AES key saved to ", keyFilePath)
 }
 
 // EncryptAESCmd encrypts a file using AES and saves the symmetric key with a UUID prefix
 func (commandHandler *AESCommandHandler) EncryptAESCmd(cmd *cobra.Command, _ []string) {
-	inputFilePath, _ := cmd.Flags().GetString("input-file")
-	outputFilePath, _ := cmd.Flags().GetString("output-file")
-	symmetricKey, _ := cmd.Flags().GetString("symmetric-key")
+	inputFilePath, err := cmd.Flags().GetString("input-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid input-file flag ", err)
+		return
+	}
+	outputFilePath, err := cmd.Flags().GetString("output-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid output-file flag ", err)
+		return
+	}
+	symmetricKey, err := cmd.Flags().GetString("symmetric-key")
+	if err != nil {
+		commandHandler.logger.Error("invalid symmetric-key flag ", err)
+		return
+	}
 
 	plainText, err := os.ReadFile(filepath.Clean(inputFilePath))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	key, err := os.ReadFile(filepath.Clean(symmetricKey))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	encryptedData, err := commandHandler.aesProcessor.Encrypt(plainText, key)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	err = os.WriteFile(outputFilePath, encryptedData, 0600)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
-	commandHandler.Logger.Info(fmt.Sprintf("Encrypted data saved to %s", outputFilePath))
+	commandHandler.logger.Info("Encrypted data saved to ", outputFilePath)
 }
 
 // DecryptAESCmd decrypts a file using AES and reads the corresponding symmetric key with a UUID prefix
 func (commandHandler *AESCommandHandler) DecryptAESCmd(cmd *cobra.Command, _ []string) {
-	inputFilePath, _ := cmd.Flags().GetString("input-file")
-	outputFilePath, _ := cmd.Flags().GetString("output-file")
-	symmetricKey, _ := cmd.Flags().GetString("symmetric-key")
+	inputFilePath, err := cmd.Flags().GetString("input-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid input-file flag ", err)
+		return
+	}
+	outputFilePath, err := cmd.Flags().GetString("output-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid output-file flag ", err)
+		return
+	}
+	symmetricKey, err := cmd.Flags().GetString("symmetric-key")
+	if err != nil {
+		commandHandler.logger.Error("invalid symmetric-key flag ", err)
+		return
+	}
 
 	key, err := os.ReadFile(filepath.Clean(symmetricKey))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	encryptedData, err := os.ReadFile(filepath.Clean(inputFilePath))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	decryptedData, err := commandHandler.aesProcessor.Decrypt(encryptedData, key)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	err = os.WriteFile(outputFilePath, decryptedData, 0600)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
-	commandHandler.Logger.Info(fmt.Sprintf("Decrypted data saved to %s.", outputFilePath))
+	commandHandler.logger.Info("Decrypted data saved to ", outputFilePath)
 }
 
 // InitAESCommands registers AES-related commands
-func InitAESCommands(rootCmd *cobra.Command) {
-	handler := NewAESCommandHandler()
+func InitAESCommands(rootCmd *cobra.Command) error {
+	handler, err := NewAESCommandHandler()
+
+	if err != nil {
+		return fmt.Errorf("failed to create AES command handler %w", err)
+	}
 
 	var generateAESKeysCmd = &cobra.Command{
 		Use:   "generate-aes-keys",
@@ -166,4 +194,6 @@ func InitAESCommands(rootCmd *cobra.Command) {
 	decryptAESFileCmd.Flags().StringP("output-file", "", "", "Path to decrypted output file")
 	decryptAESFileCmd.Flags().StringP("symmetric-key", "", "", "Path to the symmetric key")
 	rootCmd.AddCommand(decryptAESFileCmd)
+
+	return nil
 }
