@@ -2,12 +2,11 @@ package commands
 
 import (
 	"crypto/elliptic"
+	"crypto_vault_service/internal/domain/crypto"
 	"crypto_vault_service/internal/infrastructure/cryptography"
-	"crypto_vault_service/internal/infrastructure/logger"
-	"crypto_vault_service/internal/infrastructure/settings"
+	"crypto_vault_service/internal/pkg/logger"
 	"encoding/hex"
 	"fmt"
-	"log"
 	"os"
 	"path/filepath"
 
@@ -17,41 +16,40 @@ import (
 
 // ECCommandHandler encapsulates logic for handling elliptic curve cryptographic operations via CLI.
 type ECCommandHandler struct {
-	ecProcessor cryptography.ECProcessor
-	Logger      logger.Logger
+	ecProcessor crypto.ECProcessor
+	logger      logger.Logger
 }
 
 // NewECCommandHandler initializes a new ECCommandHandler with logging and an EC processor.
-// It panics if any setup step fails.
-func NewECCommandHandler() *ECCommandHandler {
-	loggerSettings := &settings.LoggerSettings{
-		LogLevel: "info",
-		LogType:  "console",
-		FilePath: "",
+func NewECCommandHandler() (*ECCommandHandler, error) {
+	loggerInstance, err := setupLogger()
+	if err != nil {
+		return nil, fmt.Errorf("failed to setup logger: %w", err)
 	}
 
-	logger, err := logger.GetLogger(loggerSettings)
+	ecProcessor, err := cryptography.NewECProcessor(loggerInstance)
 	if err != nil {
-		log.Panicf("Error creating logger: %v", err)
-		return nil
-	}
-
-	ecProcessor, err := cryptography.NewECProcessor(logger)
-	if err != nil {
-		log.Panicf("%v\n", err)
-		return nil
+		return nil, fmt.Errorf("failed to create EC processor: %w", err)
 	}
 
 	return &ECCommandHandler{
 		ecProcessor: ecProcessor,
-		Logger:      logger,
-	}
+		logger:      loggerInstance,
+	}, nil
 }
 
 // GenerateECKeysCmd generates EC key pairs and persists those in a selected directory
 func (commandHandler *ECCommandHandler) GenerateECKeysCmd(cmd *cobra.Command, _ []string) {
-	keySize, _ := cmd.Flags().GetInt("key-size")
-	keyDir, _ := cmd.Flags().GetString("key-dir")
+	keySize, err := cmd.Flags().GetInt("key-size")
+	if err != nil {
+		commandHandler.logger.Error("invalid key-size flag ", err)
+		return
+	}
+	keyDir, err := cmd.Flags().GetString("key-dir")
+	if err != nil {
+		commandHandler.logger.Error("invalid key-dir flag ", err)
+		return
+	}
 
 	uniqueID := uuid.New()
 
@@ -66,110 +64,135 @@ func (commandHandler *ECCommandHandler) GenerateECKeysCmd(cmd *cobra.Command, _ 
 	case 521:
 		curve = elliptic.P521()
 	default:
-		commandHandler.Logger.Error("key size %v not supported", keySize)
+		commandHandler.logger.Error("key size %v not supported", keySize)
 		return
 	}
 
 	privateKey, publicKey, err := commandHandler.ecProcessor.GenerateKeys(curve)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	privateKeyFilePath := fmt.Sprintf("%s/%s-private-key.pem", keyDir, uniqueID.String())
 	err = commandHandler.ecProcessor.SavePrivateKeyToFile(privateKey, privateKeyFilePath)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	publicKeyFilePath := fmt.Sprintf("%s/%s-public-key.pem", keyDir, uniqueID.String())
 	err = commandHandler.ecProcessor.SavePublicKeyToFile(publicKey, publicKeyFilePath)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 }
 
 // SignECCCmd signs the contents of a file with ECDSA
 func (commandHandler *ECCommandHandler) SignECCCmd(cmd *cobra.Command, _ []string) {
-	inputFilePath, _ := cmd.Flags().GetString("input-file")
-	privateKeyFilePath, _ := cmd.Flags().GetString("private-key")
-	signatureFilePath, _ := cmd.Flags().GetString("output-file")
+	inputFilePath, err := cmd.Flags().GetString("input-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid input-file flag ", err)
+		return
+	}
+	privateKeyFilePath, err := cmd.Flags().GetString("private-key")
+	if err != nil {
+		commandHandler.logger.Error("invalid private-key flag ", err)
+		return
+	}
+	signatureFilePath, err := cmd.Flags().GetString("output-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid output-file flag ", err)
+		return
+	}
 
 	fileContent, err := os.ReadFile(filepath.Clean(inputFilePath))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	privateKey, err := commandHandler.ecProcessor.ReadPrivateKey(privateKeyFilePath, elliptic.P256())
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	signature, err := commandHandler.ecProcessor.Sign(fileContent, privateKey)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	err = commandHandler.ecProcessor.SaveSignatureToFile(signatureFilePath, signature)
 	if err != nil {
-		if err != nil {
-			commandHandler.Logger.Error(fmt.Sprintf("%v", err))
-			return
-		}
+		commandHandler.logger.Error(err)
+		return
 	}
 }
 
 // VerifyECCCmd verifies the signature of a file's content using ECDSA
 func (commandHandler *ECCommandHandler) VerifyECCCmd(cmd *cobra.Command, _ []string) {
-	inputFilePath, _ := cmd.Flags().GetString("input-file")
-	publicKeyPath, _ := cmd.Flags().GetString("public-key")
-	signatureFile, _ := cmd.Flags().GetString("signature-file")
+	inputFilePath, err := cmd.Flags().GetString("input-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid input-file flag ", err)
+		return
+	}
+	publicKeyPath, err := cmd.Flags().GetString("public-key")
+	if err != nil {
+		commandHandler.logger.Error("invalid public-key flag ", err)
+		return
+	}
+	signatureFile, err := cmd.Flags().GetString("signature-file")
+	if err != nil {
+		commandHandler.logger.Error("invalid signature-file flag ", err)
+		return
+	}
 
 	publicKey, err := commandHandler.ecProcessor.ReadPublicKey(publicKeyPath, elliptic.P256())
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	fileContent, err := os.ReadFile(filepath.Clean(inputFilePath))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	signatureHex, err := os.ReadFile(filepath.Clean(signatureFile))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	signature, err := hex.DecodeString(string(signatureHex))
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	valid, err := commandHandler.ecProcessor.Verify(fileContent, signature, publicKey)
 	if err != nil {
-		commandHandler.Logger.Error(fmt.Sprintf("%v", err))
+		commandHandler.logger.Error(err)
 		return
 	}
 
 	if valid {
-		commandHandler.Logger.Info(fmt.Sprintf("Signature valid for %s", inputFilePath))
+		commandHandler.logger.Info("Signature valid for ", inputFilePath)
 	} else {
-		commandHandler.Logger.Info(fmt.Sprintf("Signature invalid for %s", inputFilePath))
+		commandHandler.logger.Info("Signature invalid for ", inputFilePath)
 	}
 }
 
 // InitECDSACommands registers EC-related commands
-func InitECDSACommands(rootCmd *cobra.Command) {
-	handler := NewECCommandHandler()
+func InitECDSACommands(rootCmd *cobra.Command) error {
+	handler, err := NewECCommandHandler()
+	if err != nil {
+		return fmt.Errorf("failed to create EC command handler %w", err)
+	}
 
 	var generateECKeysCmd = &cobra.Command{
 		Use:   "generate-ecc-keys",
@@ -199,4 +222,6 @@ func InitECDSACommands(rootCmd *cobra.Command) {
 	verifyECCSignatureCmd.Flags().StringP("public-key", "", "", "Path to ECC public key")
 	verifyECCSignatureCmd.Flags().StringP("signature-file", "", "", "Path to signature input file")
 	rootCmd.AddCommand(verifyECCSignatureCmd)
+
+	return nil
 }
