@@ -1,2 +1,2 @@
-// Package v1 provides gRPC services for managing blobs and cryptographic keys
+// Package v1 provides gRPC APIs for managing blobs and cryptographic keys
 package v1
