@@ -1,2 +1,2 @@
-// Package v1 provides HTTP handlers and services for managing blobs and cryptographic keys
+// Package v1 provides HTTP APIs for managing blobs and cryptographic keys
 package v1
