@@ -1,2 +1,2 @@
-// Package blobs defines the core interfaces for managing blobs
+// Package blobs defines the core interfaces and structures for managing blobs
 package blobs
