@@ -1,2 +1,3 @@
-// Package crypto ...
+// Package crypto defines the core interfaces and structures for managing and performing operations on cryptographic materials,
+// such as symmetric and asymmetric keys, including key generation, encryption, decryption, signing, verification, and PKCS#11 operations.
 package crypto
