@@ -1,2 +1,2 @@
-// Package keys defines the core interfaces for managing cryptographic keys.
+// Package keys defines the core interfaces and structures for managing cryptographic keys.
 package keys
