@@ -1,7 +1,7 @@
 package keys
 
 import (
-	"crypto_vault_service/internal/domain/validators"
+	"crypto_vault_service/internal/pkg/validators"
 	"errors"
 	"fmt"
 	"time"
@@ -9,11 +9,11 @@ import (
 	"github.com/go-playground/validator/v10"
 )
 
-// CryptoKeyMeta represents the encryption key entity
+// CryptoKeyMeta entity
 type CryptoKeyMeta struct {
 	ID              string    `gorm:"primaryKey" validate:"required,uuid4"`
 	KeyPairID       string    `gorm:"index" validate:"required,uuid4"`
-	Algorithm       string    `validate:"omitempty,oneof=AES RSA EC"`
+	Algorithm       string    `validate:"omitempty,oneof=AES RSA ECDSA"`
 	KeySize         uint32    `json:"key_size" validate:"omitempty,keySizeValidation"`
 	Type            string    `validate:"omitempty,oneof=private public symmetric"`
 	DateTimeCreated time.Time `validate:"required"`
