@@ -7,30 +7,24 @@ import (
 	"context"
 	"testing"
 
-	"crypto_vault_service/internal/infrastructure/logger"
-	"crypto_vault_service/internal/infrastructure/settings"
-	"crypto_vault_service/test/testutils"
+	"crypto_vault_service/internal/domain/blobs"
+	"crypto_vault_service/internal/pkg/config"
+	pkgTesting "crypto_vault_service/internal/pkg/testing"
 
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-// AzureBlobConnectorTest encapsulates common logic for tests
 type AzureBlobConnectorTest struct {
-	blobConnector BlobConnector
+	blobConnector blobs.BlobConnector
 }
 
-func NewAzureBlobConnectorTest(t *testing.T, cloudProvider, connectionString string, containerName string) *AzureBlobConnectorTest {
+func NewAzureBlobConnectorTest(t *testing.T, cloudProvider, connectionString, containerName string) *AzureBlobConnectorTest {
+	t.Helper()
+	logger := pkgTesting.SetupTestLogger(t)
 
-	loggerSettings := &settings.LoggerSettings{
-		LogLevel: "info",
-		LogType:  "console",
-		FilePath: "",
-	}
-	logger, err := logger.GetLogger(loggerSettings)
-	require.NoError(t, err)
-	blobConnectorSettings := &settings.BlobConnectorSettings{
+	blobConnectorSettings := &config.BlobConnectorSettings{
 		CloudProvider:    cloudProvider,
 		ConnectionString: connectionString,
 		ContainerName:    containerName,
@@ -46,21 +40,17 @@ func NewAzureBlobConnectorTest(t *testing.T, cloudProvider, connectionString str
 }
 
 func TestAzureBlobConnector_Upload(t *testing.T) {
-
-	abct := NewAzureBlobConnectorTest(t, "azure", "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;", "testblobs")
+	abct := NewAzureBlobConnectorTest(t, TestCloudProvider, TestConnectionString, TestContainerName)
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
-	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := pkgTesting.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userID := uuid.New().String()
-
-	var encryptionKeyID *string = nil
-	var signKeyID *string = nil
 	ctx := context.Background()
 
-	blobs, err := abct.blobConnector.Upload(ctx, form, userID, encryptionKeyID, signKeyID)
+	blobs, err := abct.blobConnector.Upload(ctx, form, userID, nil, nil)
 	require.NoError(t, err)
 
 	require.Len(t, blobs, 1)
@@ -69,55 +59,86 @@ func TestAzureBlobConnector_Upload(t *testing.T) {
 	assert.Equal(t, testFileName, blob.Name)
 	assert.Equal(t, int64(len(testFileContent)), blob.Size)
 	assert.Equal(t, ".txt", blob.Type)
+	assert.Nil(t, blob.EncryptionKeyID)
+	assert.Nil(t, blob.SignKeyID)
+
+	err = abct.blobConnector.Delete(ctx, blob.ID, blob.Name)
+	require.NoError(t, err)
+}
+
+func TestAzureBlobConnector_Upload_WithEncryptionAndSignKeys(t *testing.T) {
+	abct := NewAzureBlobConnectorTest(t, TestCloudProvider, TestConnectionString, TestContainerName)
+
+	testFileContent := []byte("encrypted content")
+	testFileName := "encrypted.txt"
+	form, err := pkgTesting.CreateTestFileAndForm(t, testFileName, testFileContent)
+	require.NoError(t, err)
+
+	userID := uuid.New().String()
+	encryptionKeyID := uuid.New().String()
+	signKeyID := uuid.New().String()
+	ctx := context.Background()
+
+	blobs, err := abct.blobConnector.Upload(ctx, form, userID, &encryptionKeyID, &signKeyID)
+	require.NoError(t, err)
+
+	require.Len(t, blobs, 1)
+	blob := blobs[0]
+	assert.NotNil(t, blob.EncryptionKeyID)
+	assert.Equal(t, encryptionKeyID, *blob.EncryptionKeyID)
+	assert.NotNil(t, blob.SignKeyID)
+	assert.Equal(t, signKeyID, *blob.SignKeyID)
 
 	err = abct.blobConnector.Delete(ctx, blob.ID, blob.Name)
 	require.NoError(t, err)
 }
 
 func TestAzureBlobConnector_Download(t *testing.T) {
-
-	abct := NewAzureBlobConnectorTest(t, "azure", "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;", "testblobs")
+	abct := NewAzureBlobConnectorTest(t, TestCloudProvider, TestConnectionString, TestContainerName)
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.pem"
-	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := pkgTesting.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userID := uuid.New().String()
-
-	var encryptionKeyID *string = nil
-	var signKeyID *string = nil
 	ctx := context.Background()
-	blobs, err := abct.blobConnector.Upload(ctx, form, userID, encryptionKeyID, signKeyID)
+
+	blobs, err := abct.blobConnector.Upload(ctx, form, userID, nil, nil)
 	require.NoError(t, err)
 
 	blob := blobs[0]
 
 	downloadedData, err := abct.blobConnector.Download(ctx, blob.ID, blob.Name)
 	require.NoError(t, err)
-
 	assert.Equal(t, testFileContent, downloadedData)
 
 	err = abct.blobConnector.Delete(ctx, blob.ID, blob.Name)
 	require.NoError(t, err)
 }
 
-func TestAzureBlobConnector_Delete(t *testing.T) {
+func TestAzureBlobConnector_Download_NotFound(t *testing.T) {
+	abct := NewAzureBlobConnectorTest(t, TestCloudProvider, TestConnectionString, TestContainerName)
 
-	abct := NewAzureBlobConnectorTest(t, "azure", "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;", "testblobs")
+	nonExistentID := uuid.New().String()
+	ctx := context.Background()
+
+	_, err := abct.blobConnector.Download(ctx, nonExistentID, "nonexistent.txt")
+	assert.Error(t, err)
+}
+
+func TestAzureBlobConnector_Delete(t *testing.T) {
+	abct := NewAzureBlobConnectorTest(t, TestCloudProvider, TestConnectionString, TestContainerName)
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.pem"
-	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := pkgTesting.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userID := uuid.New().String()
-
-	var encryptionKeyID *string = nil
-	var signKeyID *string = nil
 	ctx := context.Background()
 
-	blobs, err := abct.blobConnector.Upload(ctx, form, userID, encryptionKeyID, signKeyID)
+	blobs, err := abct.blobConnector.Upload(ctx, form, userID, nil, nil)
 	require.NoError(t, err)
 
 	blob := blobs[0]
@@ -126,5 +147,19 @@ func TestAzureBlobConnector_Delete(t *testing.T) {
 	require.NoError(t, err)
 
 	_, err = abct.blobConnector.Download(ctx, blob.ID, blob.Name)
+	assert.Error(t, err)
+}
+
+func TestNewAzureBlobConnector_InvalidSettings(t *testing.T) {
+	logger := pkgTesting.SetupTestLogger(t)
+	ctx := context.Background()
+
+	invalidSettings := &config.BlobConnectorSettings{
+		CloudProvider:    "",
+		ConnectionString: "",
+		ContainerName:    "",
+	}
+
+	_, err := NewAzureBlobConnector(ctx, invalidSettings, logger)
 	assert.Error(t, err)
 }
