@@ -5,7 +5,8 @@ import (
 	"crypto/elliptic"
 	"crypto/rand"
 	"crypto/sha256"
-	"crypto_vault_service/internal/infrastructure/logger"
+	"crypto_vault_service/internal/domain/crypto"
+	"crypto_vault_service/internal/pkg/logger"
 	"encoding/hex"
 	"encoding/pem"
 	"fmt"
@@ -15,25 +16,13 @@ import (
 	"path/filepath"
 )
 
-// ECProcessor Interface
-type ECProcessor interface {
-	GenerateKeys(curve elliptic.Curve) (*ecdsa.PrivateKey, *ecdsa.PublicKey, error)
-	Sign(message []byte, privateKey *ecdsa.PrivateKey) ([]byte, error)
-	Verify(message, signature []byte, publicKey *ecdsa.PublicKey) (bool, error)
-	SaveSignatureToFile(filename string, data []byte) error
-	SavePrivateKeyToFile(privateKey *ecdsa.PrivateKey, filename string) error
-	SavePublicKeyToFile(publicKey *ecdsa.PublicKey, filename string) error
-	ReadPrivateKey(privateKeyPath string, curve elliptic.Curve) (*ecdsa.PrivateKey, error)
-	ReadPublicKey(publicKeyPath string, curve elliptic.Curve) (*ecdsa.PublicKey, error)
-}
-
 // ecProcessor struct that implements the ECProcessor interface
 type ecProcessor struct {
 	logger logger.Logger
 }
 
 // NewECProcessor creates and returns a new instance of ecProcessor
-func NewECProcessor(logger logger.Logger) (ECProcessor, error) {
+func NewECProcessor(logger logger.Logger) (crypto.ECProcessor, error) {
 	return &ecProcessor{
 		logger: logger,
 	}, nil
@@ -126,7 +115,7 @@ func (e *ecProcessor) SavePrivateKeyToFile(privateKey *ecdsa.PrivateKey, filenam
 		return fmt.Errorf("failed to encode private key: %w", err)
 	}
 
-	e.logger.Info(fmt.Sprintf("Saved EC private key %s", filename))
+	e.logger.Info("Saved EC private key ", filename)
 	return nil
 }
 
@@ -156,7 +145,7 @@ func (e *ecProcessor) SavePublicKeyToFile(publicKey *ecdsa.PublicKey, filename s
 	if err != nil {
 		return fmt.Errorf("failed to encode public key: %w", err)
 	}
-	e.logger.Info(fmt.Sprintf("Saved EC public key %s", filename))
+	e.logger.Info("Saved EC public key ", filename)
 
 	return nil
 }
@@ -168,7 +157,7 @@ func (e *ecProcessor) SaveSignatureToFile(filename string, data []byte) error {
 	if err != nil {
 		return fmt.Errorf("failed to write data to file %s: %w", filename, err)
 	}
-	e.logger.Info(fmt.Sprintf("Saved signature file %s", filename))
+	e.logger.Info("Saved signature file ", filename)
 	return nil
 }
 
